+++ CHANGELOG.md
@@ -2,6 +2,7 @@ unreleased
 
 - added dummy classifier
 - add checks for matching categorical maps between train and predict
+- added titanic toy data
 
 
 0.7.2